Initial values are now used.
When a Program is compiled, any variable declarations that were
encountered during compilation will have their initial value stored
inside the Program.

If VariableStorage.GetValue returns null, the virtual machine will
consult this list of initial values, to see if a value is available
there. If it can't, an exception will be thrown, because the value of
the variable can't be determined.

(Variables that were declared externally are required to have their
initial value always available from the VariableStorage.)