Rephrased the declare statement section.
As it previously existed while it was a valid approach and technically what we do in Yarn Spinner it was very rigid and requires a two pass approach.
There is no reason that a fully type inferred version of Yarn doing this all line by line at runtime couldn't exist (assuming people use variables) properly and this would require no declarations.

The errors still exist for using variables that don't match their type.