Registered visited functions with library.
The two relevant functions, visited and visited_count are now registered with the library created by the dialogue object.
This probably shouldn't go in the dialogue itself but I suppose it isn't a huge issue.

Moved the GenerateUniquieVisitedVariableForNode method from the compiler to the library.  This is because it needs to be used both at compile time and runtime.
Not sure if that is the best spot but for now is fine.