Visit tracking now works without jumps
Previously tracking opcode were only generated when you jumped out of a node that needed to be tracked.
This meant that if you just completed a node that had no jumps it was never flagged as visited upon later runs.
Technically could be smarter and work out if a jump happens and only emit codes then blah blah but thats a lot of work and this extra increment won't happen anyways if there is a jump so its fine.