Inferred function return types are saved.
This means that not only now does the typechecker infer function returns based on more data but it also saves that inference back into the derived declarations.
This means later uses of the function can be used to determine types or inconsistent uses of the function can be detected.

As an example if you have something like the following:
<<declare $cats = 0>>
<<set $cats = catfunc()>>
<<declare $dogs = "woof">>
<<set $dogs = catfunc()>>

This will trip a diagnostic error that you are trying to set $dogs to a Number because that is what catfunc() was determined to be earlier on.
This will change depending on the ordering of files and lines in said file but it will detect these changing uses of a function now.