hack to prevent crash
So the previous code uses a First call on attributes like YarnCommand which meant if you have two or more attributes on a method and the first isnt the YarnCommand it crashes.
That is gone now but I don't like how we handle this and we should change it later.