Better escape handling
Previously if the first character of a line was the \ character it got skipped and was lexed as text instead of being flagged as being necessary to escape the next character.
This commit fixes this.

Technically this mean Yarn that wasn't actually trying to escape characters but just happened to have used the \ character as the first character did work but now will not and will need to be escaped as \\.
Considering that was a bug this is probably ok.

Fixes #324
Fixes #309