Mostly automated code cleanup.
Was annoyed by the sheer volume of warnings we were getting which made finding specific ones hard.
Ran through the warnings automatically to fix the majority but it seemed to time out so I had to do a decent chunk by hand.
So things likely got missed.

Almost everything left is a missing XML comment but there are a few I wasn't entirely sure of the best fix so left as is.

Globally suppressed the following warnings:
- 1805: initialising with defaults (this is a bad warning IMO)
- 1819: Properties shouldn't return arrays
- 1002: Shouldn't expose lists
for 1819 and 1002 I partially agree with MS but it isn't like these are huge issues and would be a big breaking code change.
If we feel we want to change these because the are globally suppressed we can just remove those lines from the editor config.

Also I added a quick copyright header to every file (again it timed out so I think it got them all...). This is something we should look at making the build process do.