Added a new method for tagging lines.
Works in an identical fashion to the existing method AddTagsToLines but returns the list of tags in addition to the modified text.
Have marked the original method as obsolete as it does still work but this new one is more flexible, expect it to go away in a later version.