Attempt to reduce spurious error messages
After type checking, if there are unresolved (or unresolvable) type
constraints, it's likely that not all of them are actually errors. This
change causes the type checker to attempt to reduce the number of
failing constraints before they get turned into error messages (and
should eliminate absurd messages like "Number is not convertible to
Number").