handled escaping characters at the start of lines.
Because of the quirks of markup it meant that it was losing the escaped character and that was causing all sorts of issues when it occurs at the start of a line.
Added a new specific rule before escaping occurs to catch this, not the best approach but it works for now.
Fixes #377